# docs/swagger.yaml
definitions:
...
  enum.UserRole:
    enum:
    - ADMIN
    - MARKETING_STAFF
    - CONTENT_STAFF
    - SALES_STAFF
    - CUSTOMER
    - BRAND_PARTNER
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleMarketingStaff
    - UserRoleContentStaff
    - UserRoleSalesStaff
    - UserRoleCustomer
    - UserRoleBrandPartner
...
  requests.UpdateAddressProfileRequest:
    properties:
      address_line_2:
        example: Apt 4B
        maxLength: 200
        type: string
      city:
        example: New York
        maxLength: 100
        minLength: 2
        type: string
      company:
        example: Acme Corp
        maxLength: 100
        type: string
      country:
        example: USA
        maxLength: 100
        minLength: 2
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        maxLength: 100
        minLength: 3
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_default:
        example: false
        type: boolean
      phone_number:
        example: "+1234567890"
        type: string
      postal_code:
        example: "10001"
        maxLength: 20
        minLength: 2
        type: string
      state:
        example: NY
        maxLength: 100
        minLength: 2
        type: string
      street:
        example: 123 Main St
        maxLength: 200
        minLength: 5
        type: string
      type:
        enum:
        - BILLING
        - SHIPPING
        example: SHIPPING
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - city
    - country
    - email
    - full_name
    - id
    - phone_number
    - postal_code
    - street
    - type
    - user_id
    type: object
...
  requests.UpdateProfileRequest:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      date_of_birth:
        example: "1990-01-01"
        type: string
      full_name:
        example: John Doe
        maxLength: 100
        minLength: 3
        type: string
      phone:
        example: "+1234567890"
        type: string
      shipping_address:
        items:
          $ref: '#/definitions/requests.UpdateAddressProfileRequest'
        type: array
      username:
        example: new_username
        maxLength: 50
        minLength: 3
        type: string
    type: object
  requests.UpdateUserRoleRequest:
    properties:
      role:
        enum:
        - ADMIN
        - MARKETING_STAFF
        - CONTENT_STAFF
        - SALES_STAFF
        - CUSTOMER
        - BRAND_PARTNER
        example: CUSTOMER
        type: string
    required:
    - role
    type: object
  requests.UpdateUserStatusRequest:
    properties:
      is_active:
        example: true
        type: boolean
    required:
    - is_active
    type: object
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
  responses.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  responses.ShippingAddressResponse:
    properties:
      address_line_2:
        example: Apt 4B
        type: string
      city:
        example: New York
        type: string
      company:
        example: Acme Corp
        type: string
      country:
        example: USA
        type: string
      created_at:
        example: "2023-12-30 15:04:05"
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_default:
        example: false
        type: boolean
      phone_number:
        example: "+1234567890"
        type: string
      postal_code:
        example: "10001"
        type: string
      state:
        example: NY
        type: string
      street:
        example: 123 Main St
        type: string
      type:
        example: SHIPPING
        type: string
      updated_at:
        example: "2023-12-30 15:04:05"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  responses.UserListResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      is_brand_account:
        example: false
        type: boolean
      number_of_sessions:
        example: 3
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/enum.UserRole'
        example: user
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
  responses.UserPaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.UserListResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/responses.Pagination'
      status:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  responses.UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      current_login_device:
        example:
        - '["Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML'
        - ' like Gecko) Chrome/58.0.3029.110 Safari/537.3"]'
        items:
          type: string
        type: array
      date_of_birth:
        example: "1990-01-01"
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      is_brand_account:
        example: false
        type: boolean
      last_login:
        example: "2023-01-01T00:00:00Z"
        type: string
      number_of_sessions:
        example: 3
        type: integer
      phone:
        example: "+1234567890"
        type: string
      role:
        allOf:
        - $ref: '#/definitions/enum.UserRole'
        example: user
      shipping_address:
        items:
          $ref: '#/definitions/responses.ShippingAddressResponse'
        type: array
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for SEP490 project with authentication,
    user management, and business features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SEP490 Core Backend API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Field to sort by
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      - description: Search term for username or email
        in: query
        name: search
        type: string
      - description: Filter by user role
        in: query
        name: role
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Filter by brand account status
        in: query
        name: is_brand_account
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/responses.UserPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Users List
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user details by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.UserResponse'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - Users
  /api/v1/users/{id}/activate-brand:
    patch:
      consumes:
      - application/json
      description: Activate a user associated with a brand (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Brand user activated successfully
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Activate Brand User
      tags:
      - Users
  /api/v1/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Update user role (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User role updated successfully
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Update User Role
      tags:
      - Users
  /api/v1/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Update user active status (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User status updated successfully
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Update User Status
      tags:
      - Users
  /api/v1/users/profile:
    get:
      consumes:
      - application/json
      description: Get current authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update current authenticated user's profile
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.UserResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"